Consolidate everything in a single index.ts
